split out library and binary code
